refactor(desktop): isolate Home value components
Keeps DashboardPage below the product file size ratchet and removes the remaining purple Home glow.

Verified: desktop/macos/test.sh passed all launcher, 376 Rust, and 338 isolated Swift suites.
