+++ desktop/macos/Desktop/Tests/DashboardCaptureStateTests.swift
@@ -60,8 +60,7 @@ final class DashboardCaptureStateTests: XCTestCase {
     XCTAssertTrue(source.contains("private func homeHubStage(askBarWidth: CGFloat, stageHeight: CGFloat) -> some View"))
     XCTAssertTrue(source.contains("HomeValueHero(snapshot: homeValueSnapshot)"))
     XCTAssertTrue(source.contains("ScrollView {"))
-    XCTAssertTrue(source.contains("private struct HomeStatRibbon: View"))
-    XCTAssertTrue(source.contains(".frame(height: 76)"))
+    XCTAssertTrue(source.contains("static let stageGlow = Color.white"))
     XCTAssertTrue(source.contains("placeholder: homeValueSnapshot.askPlaceholder"))
     XCTAssertFalse(source.contains("private var homeHubWordmark: some View"))
     XCTAssertFalse(source.contains(".frame(width: 304)"))
